+++ build.sh
@@ -3,6 +3,7 @@
 # Please check pnda-build/ for the build products
 
 VERSION=${1}
+HADOOP_FLAVOR=${2}
 
 function error {
     echo "Not Found"
@@ -28,7 +29,7 @@ cd src/main/resources
 PYLINTOUT=$(find . -type f -name '*.py' | grep -vi __init__ | xargs pylint)
 SCORE=$(echo ${PYLINTOUT} | grep -Po '(?<=rated at ).*?(?=/10)')
 echo ${SCORE}
-if [[ $(bc <<< "${SCORE} > 8") == 0 ]]; then
+if [[ $(bc <<< "${SCORE} > 9") == 0 ]]; then
     code_quality_error "${PYLINTOUT}"
 fi
 
@@ -43,9 +44,9 @@ cd ${BASE}
 # Build
 mkdir -p pnda-build
 mvn versions:set -DnewVersion=${VERSION}
-mvn clean package
-mv target/platform-testing-cdh-${VERSION}.tar.gz pnda-build/
+mvn clean package -P${HADOOP_FLAVOR}
+mv target/platform-testing-${HADOOP_FLAVOR}-${VERSION}.tar.gz pnda-build/
 mv target/platform-testing-general-${VERSION}.tar.gz pnda-build/
-sha512sum pnda-build/platform-testing-cdh-${VERSION}.tar.gz > pnda-build/platform-testing-cdh-${VERSION}.tar.gz.sha512.txt
+sha512sum pnda-build/platform-testing-${HADOOP_FLAVOR}-${VERSION}.tar.gz > pnda-build/platform-testing-${HADOOP_FLAVOR}-${VERSION}.tar.gz.sha512.txt
 sha512sum pnda-build/platform-testing-general-${VERSION}.tar.gz > pnda-build/platform-testing-general-${VERSION}.tar.gz.sha512.txt