+++ src/main/resources/plugins/kafka/TestbotPlugin.py
@@ -26,29 +26,37 @@ import time
 import argparse
 import sys
 import os
+import math
 import logging
-
+import json
+from decimal import Decimal
 import requests
-
 from prettytable import PrettyTable
-
 from plugins.common.zkclient import ZkClient, ZkError
 from plugins.kafka.prod2cons import Prod2Cons
 from plugins.common.defcom import MonitorSummary, PartitionState, TestbotResult
 from plugins.common.defcom import ZkNodesHealth, ZkNode, KkBroker
-
 from pnda_plugin import PndaPlugin
 from pnda_plugin import Event
 from pnda_plugin import MonitorStatus
-from decimal import Decimal
 
 sys.path.insert(0, '../..')
 
-TestbotPlugin = lambda: KafkaWhitebox() # pylint: disable=invalid-name
-
+TESTBOTPLUGIN = lambda: KafkaWhitebox()
 TIMESTAMP_MILLIS = lambda: int(time.time() * 1000)
 HERE = os.path.abspath(os.path.dirname(__file__))
-LOGGER = logging.getLogger("TestbotPlugin")
+LOGGER = logging.getLogger("TESTBOTPLUGIN")
+NBTEST = 10
+
+def get_broker_by_id(brokers, search):
+    '''
+    Get broker by id
+    '''
+    for broker in brokers.list:
+        if broker.id == search:
+            return KkBroker(
+                search, broker.host, broker.port, broker.jmx_port, broker.alive)
+    return None
 
 class ProcessorError(Exception):
     '''
@@ -68,19 +76,12 @@ class KafkaWhitebox(PndaPlugin):
 
     def __init__(self):
         self.broker_list = []
-        self.zk_list = []
-        self.postjson = False
         self.display = False
         self.results = []
         self.topic_list = []
         self.prod2cons = False
-        self.runtesttopic = "avro.internal.testbot"
-        self.avro_schema = "%s/%s" % (HERE, "dataplatform-raw.avsc")
-        self.runnbtest = 10
-        self.consumer_timeout = 1  # max number of second to wait for
         self.whitebox_error_code = -1
         self.activecontrollercount = -1
-        self.unclean_threshold = 0.0002
 
     def read_args(self, args):
         '''
@@ -92,6 +93,8 @@ class KafkaWhitebox(PndaPlugin):
             usage='%(prog)s [options]',
             description='Show state of Zk-Kafka cluster',
             add_help=False)
+        parser.add_argument('--jmxproxy', default='127.0.0.1:8000',
+                            help='host:port of the jmxproxy to use')
         parser.add_argument('--brokerlist', default='localhost:9092',
                             help='comma separated host:port pairs, each corresponding to ' + \
           'a kafka broker (default: localhost:9092)')
@@ -106,19 +109,14 @@ class KafkaWhitebox(PndaPlugin):
         '''
         Get brokertopicmetrics
         '''
-        self.results.append(Event(TIMESTAMP_MILLIS(),
-                                  'kafka',
-                                  'kafka.brokers.%d.topics.%s.health' %
-                                  (broker_id, topic), [], "OK")
-                            )
         for jmx_path_name in ["BytesInPerSec", "BytesOutPerSec", \
                               "MessagesInPerSec"]:
             for jmx_data in ["RateUnit", "OneMinuteRate", \
                              "EventType", "Count", "FifteenMinuteRate",
                              "FiveMinuteRate", "MeanRate"]:
-                url_jmxproxy = ("http://127.0.0.1:8000/jmxproxy/%s/"
+                url_jmxproxy = ("http://%s/jmxproxy/%s/"
                                 "kafka.server:type=BrokerTopicMetrics,"
-                                "name=%s,topic=%s/%s") % (host, jmx_path_name, topic, jmx_data)
+                                "name=%s,topic=%s/%s") % (self.jmxproxy, host, jmx_path_name, topic, jmx_data)
 
                 response = requests.get(url_jmxproxy)
                 if response.status_code == 200:
@@ -145,71 +143,13 @@ class KafkaWhitebox(PndaPlugin):
 
         return None
 
-    def get_operatingsysteminfo(self, host, broker_id):
-        '''
-        Get operatingsysteminfo
-        '''
-        for jmx_data in ["OpenFileDescriptorCount",
-                         "CommittedVirtualMemorySize",
-                         "FreePhysicalMemorySize",
-                         "SystemLoadAverage",
-                         "Arch",
-                         "ProcessCpuLoad",
-                         "FreeSwapSpaceSize",
-                         "TotalPhysicalMemorySize",
-                         "Name",
-                         "ObjectName",
-                         "TotalSwapSpaceSize",
-                         "ProcessCpuTime",
-                         "MaxFileDescriptorCount",
-                         "SystemCpuLoad",
-                         "Version",
-                         "AvailableProcessors"]:
-            url_jmxproxy = "http://127.0.0.1:8000/jmxproxy/" + \
-              "%s/java.lang:type=OperatingSystem/%s" % (host, jmx_data)
-
-            response = requests.get(url_jmxproxy)
-            if response.status_code == 200:
-                LOGGER.debug("Getting %s fo %s", response.text, url_jmxproxy)
-                self.results.append(Event(TIMESTAMP_MILLIS(),
-                                          'kafka',
-                                          'kafka.brokers.%d.system.%s' %
-                                          (broker_id, jmx_data), [], response.text))
-            else:
-                LOGGER.error("ERROR for url_jmxproxy: %s", url_jmxproxy)
-
-        return None
-
-    def get_underreplicatedpartitions(self, host, broker_id):
-        '''
-        Get underreplicatedpartitions
-        '''
-        url_jmxproxy = ("http://127.0.0.1:8000/jmxproxy/%s/"
-                        "kafka.server:type=ReplicaManager,"
-                        "name=UnderReplicatedPartitions/Value") % host
-
-        response = requests.get(url_jmxproxy)
-        if response.status_code == 200:
-            LOGGER.debug("Getting %s fo %s", response.text, url_jmxproxy)
-            self.results.append(Event(TIMESTAMP_MILLIS(),
-                                      'kafka',
-                                      'kafka.brokers.%d.UnderReplicatedPartitions' %
-                                      broker_id, [], response.text))
-            if response.text != "0":
-                self.whitebox_error_code = 101
-
-        else:
-            LOGGER.error("ERROR for url_jmxproxy: %s", url_jmxproxy)
-
-        return None
-
     def get_activecontrollercount(self, host, broker_id):
         '''
         Get activecontrollercount
         '''
-        url_jmxproxy = ("http://127.0.0.1:8000/jmxproxy/%s/"
+        url_jmxproxy = ("http://%s/jmxproxy/%s/"
                         "kafka.controller:type=KafkaController,"
-                        "name=ActiveControllerCount/Value") % host
+                        "name=ActiveControllerCount/Value") % (self.jmxproxy, host)
 
         response = requests.get(url_jmxproxy)
         if response.status_code == 200:
@@ -228,45 +168,6 @@ class KafkaWhitebox(PndaPlugin):
 
         return None
 
-    def get_leaderelectionrateandtimems(self, host, broker_id):
-        '''
-        Get leaderelectionrateandtimems
-        '''
-        for jmx_data in ["StdDev",
-                         "75thPercentile",
-                         "Mean",
-                         "LatencyUnit",
-                         "RateUnit",
-                         "98thPercentile",
-                         "95thPercentile",
-                         "99thPercentile",
-                         "EventType",
-                         "Max",
-                         "Count",
-                         "FiveMinuteRate",
-                         "MeanRate",
-                         "50thPercentile",
-                         "OneMinuteRate",
-                         "Min",
-                         "999thPercentile",
-                         "FifteenMinuteRate"]:
-            url_jmxproxy = ("http://127.0.0.1:8000/jmxproxy/%s/"
-                            "kafka.controller:type=ControllerStats,"
-                            "name=LeaderElectionRateAndTimeMs/%s") % (host, jmx_data)
-
-            response = requests.get(url_jmxproxy)
-            if response.status_code == 200:
-                LOGGER.debug("Getting %s fo %s", response.text, url_jmxproxy)
-                self.results.append(Event(TIMESTAMP_MILLIS(),
-                                          'kafka',
-                                          'kafka.brokers.%d.controllerstats.LeaderElectionRateAndTimeMs.%s' %
-                                          (broker_id, jmx_data), [], response.text))
-
-            else:
-                LOGGER.error("ERROR for url_jmxproxy: %s", url_jmxproxy)
-
-        return None
-
     def get_uncleanleaderelections(self, host, broker_id):
         '''
         Get UncleanLeaderElectionsPerSec
@@ -280,9 +181,9 @@ class KafkaWhitebox(PndaPlugin):
                          "FifteenMinuteRate",
                          "FiveMinuteRate",
                          "MeanRate"]:
-            url_jmxproxy = ("http://127.0.0.1:8000/jmxproxy/%s/"
+            url_jmxproxy = ("http://%s/jmxproxy/%s/"
                             "kafka.controller:type=ControllerStats,"
-                            "name=UncleanLeaderElectionsPerSec/%s") % (host, jmx_data)
+                            "name=UncleanLeaderElectionsPerSec/%s") % (self.jmxproxy, host, jmx_data)
 
             response = requests.get(url_jmxproxy)
             if response.status_code == 200:
@@ -294,27 +195,18 @@ class KafkaWhitebox(PndaPlugin):
                                           (broker_id, jmx_data), [], response.text))
 
                 if jmx_data == "Count":
-                  unclean_count = int(response.text)
+                    unclean_count = int(response.text)
                 elif jmx_data == "FifteenMinuteRate":
-                  unclean_rate = Decimal(response.text)
+                    unclean_rate = Decimal(response.text)
 
             else:
                 LOGGER.error("ERROR for url_jmxproxy: %s", url_jmxproxy)
         if unclean_count is not None and unclean_rate is not None:
-          if unclean_rate > self.unclean_threshold:
-            LOGGER.debug("broker %d threshold is %f and current rate is %f" % (broker_id,self.unclean_threshold, unclean_rate))
-            self.whitebox_error_code = 104
+            if unclean_rate > 0.0002:
+                LOGGER.debug("broker %d threshold is %f and current rate is %f",
+                             broker_id, 0.0002, unclean_rate)
+                self.whitebox_error_code = 104
 
-        return None
-
-    def get_broker_by_id(self, brokers, search):
-        '''
-        Get broker by id
-        '''
-        for broker in brokers.list:
-            if broker.id == search:
-                return KkBroker(
-                    search, broker.host, broker.port, broker.jmx_port, broker.alive)
         return None
 
     def process(self, zknodes, gbrokers, partitions):
@@ -330,9 +222,9 @@ class KafkaWhitebox(PndaPlugin):
             if obj.partitions["valid"] is True:
                 for parts in parts_object:
                     # Get the partition leader
-                    for part, partinfo in parts.iteritems():
+                    for part, partinfo in parts.items():
                         leader_read = partinfo['leader']
-                        broker = self.get_broker_by_id(
+                        broker = get_broker_by_id(
                             gbrokers, '%d' % leader_read)
 
                         if broker is not None:
@@ -419,18 +311,18 @@ class KafkaWhitebox(PndaPlugin):
                     bconnect += ","
                 bconnect += "%s:%d" % (host, port)
                 try:
-                    client = ZkClient(host, port)
-                    if client.ping():
-                        node_list.append(ZkNode(host, port, True))
-                        zok += 1
-                    else:
-                        if berror != "":
-                            berror += ","
-                        berror += "%s:%d" % (host, port)
-                        node_list.append(ZkNode(host, port, False))
-                        zko += 1
-                        LOGGER.error(
-                            "Zookeeper node unreachable (%s:%d)", host, port)
+                    with ZkClient(host, port) as client:
+                        if client.ping():
+                            node_list.append(ZkNode(host, port, True))
+                            zok += 1
+                        else:
+                            if berror != "":
+                                berror += ","
+                            berror += "%s:%d" % (host, port)
+                            node_list.append(ZkNode(host, port, False))
+                            zko += 1
+                            LOGGER.error(
+                                "Zookeeper node unreachable (%s:%d)", host, port)
                 except ZkError:
                     LOGGER.error(
                         "Zookeeper node unreachable (%s:%d)", host, port)
@@ -449,15 +341,19 @@ class KafkaWhitebox(PndaPlugin):
         analyse_status = MonitorStatus["green"]
         analyse_causes = []
         analyse_metric = 'kafka.health'
+        zk_majority = int(math.ceil(float(len(zk_data.list_zk.split(",")))/2))
 
-        if zk_data and len(zk_data.list_zk_ko) > 0:
-            LOGGER.error(
-                "analyse_results : at least one zookeeper node failed")
-            analyse_status = MonitorStatus["red"]
-            analyse_causes.append(
-                "zookeeper node(s) unreachable (%s)" % zk_data.list_zk_ko)
+        if zk_data and zk_data.list_zk_ko:
+            if zk_data.num_zk_ok >= zk_majority:
+                LOGGER.warn("analyse_results : at least one zookeeper node failed")
+                analyse_status = MonitorStatus["amber"]
+                analyse_causes.append("zookeeper node(s) unreachable (%s)" % zk_data.list_zk_ko)
+            else:
+                LOGGER.error("analyse_results : at least one zookeeper node failed")
+                analyse_status = MonitorStatus["red"]
+                analyse_causes.append("zookeeper node(s) unreachable (%s)" % zk_data.list_zk_ko)
 
-        if zk_data and len(zk_data.list_brokers_ko) > 0:
+        if zk_data and zk_data.list_brokers_ko:
             LOGGER.error("analyse_results : at least one broker failed")
             analyse_status = MonitorStatus["red"]
             analyse_causes.append("broker(s) unreachable (%s)" %
@@ -487,13 +383,14 @@ class KafkaWhitebox(PndaPlugin):
         if self.whitebox_error_code != -1:
 
             if self.whitebox_error_code == 101:
-                LOGGER.warn("analyse_results : UnderReplicatedPartitions should be 0")
+                LOGGER.warn("UnderReplicatedPartitions should be 0")
                 if analyse_status != MonitorStatus["red"]:
                     analyse_status = MonitorStatus["amber"]
                 analyse_causes.append(
                     "UnderReplicatedPartitions should be 0")
             elif self.whitebox_error_code == 102:
-                LOGGER.warn("analyse_results : ActiveControllerCount only one broker in the cluster should have 1")
+                LOGGER.warn(
+                    "ActiveControllerCount only one broker in the cluster should have 1")
                 if analyse_status != MonitorStatus["red"]:
                     analyse_status = MonitorStatus["amber"]
                 analyse_causes.append(
@@ -514,14 +411,36 @@ class KafkaWhitebox(PndaPlugin):
         Process the brokers
         '''
         # todo see brokerID
-        for broker_index in xrange(1, len(self.broker_list) + 1):
+        jmx_config = json.load(open("%s/%s" % (HERE, "jmx_config.json")))
+        self.results.append(Event(TIMESTAMP_MILLIS(),
+                                  'kafka',
+                                  'kafka.available.topics',
+                                  [],
+                                  json.dumps(self.topic_list)))
+
+        for broker_index in range(1, len(self.broker_list) + 1):
             broker = self.broker_list[broker_index - 1]
             for topic in self.topic_list:
                 self.get_brokertopicmetrics(broker, topic, broker_index)
-            self.get_operatingsysteminfo(broker, broker_index)
-            self.get_underreplicatedpartitions(broker, broker_index)
+                for jmx_data in jmx_config["mBeans"]:
+                    url_jmxproxy = "http://%s/jmxproxy/%s/%s" % (self.jmxproxy, broker, jmx_data["path"])
+                    LOGGER.info(url_jmxproxy)
+                    response = requests.get(url_jmxproxy)
+                    if response.status_code == 200:
+                        LOGGER.debug("Getting %s fo %s", response.text, url_jmxproxy)
+                        self.results.append(Event(TIMESTAMP_MILLIS(),
+                                                  'kafka',
+                                                  'kafka.brokers.%d.%s' %
+                                                  (broker_index, jmx_data["label"]),
+                                                  [],
+                                                  response.text))
+                        if 'expect_value' in jmx_data and (int(response.text) != jmx_data["expect_value"]):
+                            self.whitebox_error_code = jmx_data["error_code"]
+
+                    else:
+                        LOGGER.error("ERROR for url_jmxproxy: %s", url_jmxproxy)
+
             self.get_activecontrollercount(broker, broker_index)
-            self.get_leaderelectionrateandtimems(broker, broker_index)
             self.get_uncleanleaderelections(broker, broker_index)
         return None
 
@@ -537,41 +456,41 @@ class KafkaWhitebox(PndaPlugin):
         table = PrettyTable(['Broker', 'Port', 'Topic', 'PartId', 'Valid'])
         table.align['broker'] = 'l'
 
-        if zk_data and len(zk_data.partitions) > 0:
+        if zk_data and zk_data.partitions:
             for part in zk_data.partitions:
                 table.add_row(
                     [part.broker, part.port, part.topic, \
                     part.partId, part.alive])
 
         if zk_data:
-            print table.get_string(sortby='Broker')
-            print
-            print 'List of brokers:            %s' % zk_data.list_brokers
-            print 'List of brokers (ko):       %s' % zk_data.list_brokers_ko
-            print 'Number of brokers (ok):     %d' % zk_data.num_brokers_ok
-            print 'Number of brokers (ko):     %d' % zk_data.num_brokers_ko
-            print 'List of zk:                 %s' % zk_data.list_zk
-            print 'List of zk (ko):            %s' % zk_data.list_zk_ko
-            print 'Number of zk nodes (ok):    %d' % zk_data.num_zk_ok
-            print 'Number of zk nodes (ko):    %d' % zk_data.num_zk_ko
-            print 'Number of partitions (ok):  %d' % zk_data.num_part_ok
-            print 'Number of partitions (ko):  %d' % zk_data.num_part_ko
-            print 'Number of partitions:       %d' % zk_data.num_partitions
-            print 'Run (total):                %d' % self.runnbtest
-            print 'Run (sent):                 %d' % test_result.sent
-            print 'Run (rcv):                  %d' % test_result.received
-            print 'Run (total):                %d' % test_result.notvalid
-            print 'Run (avg ms):               %d' % test_result.avg_ms
+            print(table.get_string(sortby='Broker'))
+            print()
+            print('List of brokers:            %s' % zk_data.list_brokers)
+            print('List of brokers (ko):       %s' % zk_data.list_brokers_ko)
+            print('Number of brokers (ok):     %d' % zk_data.num_brokers_ok)
+            print('Number of brokers (ko):     %d' % zk_data.num_brokers_ko)
+            print('List of zk:                 %s' % zk_data.list_zk)
+            print('List of zk (ko):            %s' % zk_data.list_zk_ko)
+            print('Number of zk nodes (ok):    %d' % zk_data.num_zk_ok)
+            print('Number of zk nodes (ko):    %d' % zk_data.num_zk_ko)
+            print('Number of partitions (ok):  %d' % zk_data.num_part_ok)
+            print('Number of partitions (ko):  %d' % zk_data.num_part_ko)
+            print('Number of partitions:       %d' % zk_data.num_partitions)
+            print('Run (total):                %d' % NBTEST)
+            print('Run (sent):                 %d' % test_result.sent)
+            print('Run (rcv):                  %d' % test_result.received)
+            print('Run (total):                %d' % test_result.notvalid)
+            print('Run (avg ms):               %d' % test_result.avg_ms)
 
-        print '-' * 50
-        print 'overall status: ',
-        print "OK" if results_summary.value == MonitorStatus["green"] else \
+        print('-' * 50)
+        print('overall status:',
+              "OK" if results_summary.value == MonitorStatus["green"] else \
               "WARN" if results_summary.value == MonitorStatus["amber"] else \
-              "ERROR"
+              "ERROR")
         if results_summary.value != MonitorStatus["green"]:
-            print 'causes:'
-            print results_summary.causes
-        print '-' * 50
+            print('causes:')
+            print(results_summary.causes)
+        print('-' * 50)
         LOGGER.debug("do_display finished")
 
     def runner(self, args, display=True):
@@ -581,14 +500,14 @@ class KafkaWhitebox(PndaPlugin):
         LOGGER.debug("runner started")
 
         plugin_args = args.split() \
-            if args is not None and (len(args.strip()) > 0) \
+            if args is not None and args.strip() \
             else ""
 
         options = self.read_args(plugin_args)
 
         self.broker_list = options.brokerlist.split(",")
-        self.zk_list = options.zkconnect.split(",")
         self.prod2cons = options.prod2cons
+        self.jmxproxy = options.jmxproxy
 
         zknodes = self.getzknodes(options.zkconnect)
         LOGGER.debug(zknodes)
@@ -599,20 +518,21 @@ class KafkaWhitebox(PndaPlugin):
             LOGGER.debug("processing %s:%d", zkn.host, zkn.port)
             if zkn.alive is True:
                 try:
-                    client = ZkClient(zkn.host, zkn.port)
-                    brokers = client.brokers()
-                    topics = client.topics()
-                    for topic in topics:
-                        if not topic.id in self.topic_list:
-                            self.topic_list.append(topic.id)
-                            LOGGER.debug(
-                                "adding %s to the topic list", topic.id)
+                    with ZkClient(zkn.host, zkn.port) as client:
+                        brokers = client.brokers()
+                        topics = client.topics()
 
-                    zk_data = self.process(zknodes, brokers, topics)
-                except ZkError, exc:
+                        for topic in topics:
+                            if not topic.id in self.topic_list:
+                                self.topic_list.append(topic.id)
+                                LOGGER.debug(
+                                    "adding %s to the topic list", topic.id)
+
+                        zk_data = self.process(zknodes, brokers, topics)
+                except ZkError as exc:
                     LOGGER.error('Failed to access Zookeeper: %s', str(exc))
                     break
-                except ProcessorError, exc:
+                except ProcessorError as exc:
                     LOGGER.error('Failed to process: %s', str(exc))
                     break
                 if prev_zk_data is not None:
@@ -620,7 +540,8 @@ class KafkaWhitebox(PndaPlugin):
                             prev_zk_data.num_partitions != zk_data.num_partitions or
                             prev_zk_data.num_part_ok != zk_data.num_part_ok or
                             prev_zk_data.num_part_ko != zk_data.num_part_ko):
-                        LOGGER.error("Inconsistency found in zk (%s,%d) tree comparison", zkn.host, zkn.port)
+                        LOGGER.error("Inconsistency found in zk (%s,%d) tree comparison",
+                                     zkn.host, zkn.port)
                     else:
                         LOGGER.debug("No inconsistency found in zk (%s,%d) tree comparison", \
                                      zkn.host, zkn.port)
@@ -631,8 +552,8 @@ class KafkaWhitebox(PndaPlugin):
                                      list_brokers_ko="",
                                      num_brokers_ok=-1,
                                      num_brokers_ko=-1,
-                                     list_zk=self.zconnect,
-                                     list_zk_ko=self.zconnect,
+                                     list_zk=zknodes,
+                                     list_zk_ko=zknodes,
                                      num_zk_ok=0,
                                      num_zk_ko=len(zknodes.list),
                                      num_part_ok=-1,
@@ -643,22 +564,21 @@ class KafkaWhitebox(PndaPlugin):
         if self.prod2cons:
             LOGGER.debug("=> E2E producer / consumer test required")
             # Now, pick up a broker and run a prod2cons test run
-            if brokers and len(brokers.connect) > 0:
+            if brokers and brokers.connect:
                 # beta1: use the first of the list
                 pairbrokers = brokers.connect.split(',')
                 shost, sport = pairbrokers[0].split(':')
                 try:
                     test_runner = Prod2Cons(shost,
                                             int(sport),
-                                            self.avro_schema,
-                                            self.runtesttopic,
-                                            self.runnbtest,
-                                            self.consumer_timeout)
-                    test_runner.consumer_reset()
-                    test_runner.prod()
+                                            "%s/%s" % (HERE, "dataplatform-raw.avsc"),
+                                            "avro.internal.testbot",
+                                            NBTEST)
+                    msgsent = test_runner.prod()
+                    LOGGER.debug("prod sent %d messages", msgsent)
                     test_result = test_runner.cons()
                 except ValueError as error:
-                    LOGGER.error("Error on Prod2Cons " + str(error))
+                    LOGGER.error("Error on Prod2Cons %s", str(error))
             else:
                 LOGGER.error("No valid broker found for running prod2cons run")